custommma: Add custom MMA for more optimizations
**SUMMARY**
This will allow us to hack into the MMA itself to improve things further, like
by having one operand pre-loaded in shared memory

This PR already delivers some speedups by:
(1) Not loading the last tensor row (which isn't used until the next iteration)
(2) Using less shared memory, as we have an upperbound on K (if K<=64, we don't
need a buffer with 96 values for the first row in shared memory, 64 are enough)

I only touched the A100 MMA (in multistage) and didn't modify mma_pipelined at this point

ghstack-source-id: 550ac3611a8e4b190bf00aee091308de2cf78080
Pull Request resolved: https://github.com/facebookresearch/xformers/pull/415

+++ xformers/components/attention/csrc/cuda/mem_eff_attention/attention_backward_generic.cu
@@ -71,19 +71,19 @@ mem_efficient_attention_backward_generic(
       at::cuda::getDeviceProperties(query.device().index());
   const int computeCapability = properties->major * 10 + properties->minor;
 
-#define DISPATCH_MAXK(func)                                          \
-  {                                                                  \
-    const auto maxK = std::max(query.size(2), value.size(2));        \
-    if (maxK <= 64) {                                                \
-      constexpr int64_t kMaxK = 64;                                  \
-      func();                                                        \
-    } else if (maxK <= 128) {                                        \
-      constexpr int64_t kMaxK = 128;                                 \
-      func();                                                        \
-    } else {                                                         \
-      constexpr int64_t kMaxK = std::numeric_limits<int64_t>::max(); \
-      func();                                                        \
-    }                                                                \
+#define DISPATCH_MAXK(func)                                   \
+  {                                                           \
+    const auto maxK = std::max(query.size(2), value.size(2)); \
+    if (maxK <= 64) {                                         \
+      constexpr int kMaxK = 64;                               \
+      func();                                                 \
+    } else if (maxK <= 128) {                                 \
+      constexpr int kMaxK = 128;                              \
+      func();                                                 \
+    } else {                                                  \
+      constexpr int kMaxK = std::numeric_limits<int>::max();  \
+      func();                                                 \
+    }                                                         \
   }
 
   DISPATCH_MAXK(([&] {
